Add Moment Video + New Recommendation System

+++ src/database/migrations/20240107143405-create-contact.js
@@ -11,7 +11,7 @@ module.exports = {
                 allowNull: false,
             },
             user_id: {
-                type: Sequelize.INTEGER(),
+                type: Sequelize.BIGINT(),
                 allowNull: false,
                 references: { model: "users", key: "id" },
                 onUpdate: "CASCADE",
